requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
plotly==5.17.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2



